Separate the gdpr validation dag from snowflake delivery
